requirements.txt
requests
python-dotenv
litellm
schedule
